Demo nlohmann JSON for save files

+++ lib/SXNGN/headers/Constants.h
@@ -5,14 +5,14 @@ namespace SXNGN {
 	const std::string BAD_STRING_RETURN = "bad_string_result";
 
 	//Tile constants
-	const int BASE_TILE_WIDTH = 16;
-	const int BASE_TILE_HEIGHT = 16;
+	const size_t BASE_TILE_WIDTH = 16;
+	const size_t BASE_TILE_HEIGHT = 16;
 
-	const int PIXELS_TO_METERS = 16;
+	const size_t PIXELS_TO_METERS = 16;
 
 
-	const int TILE_WIDTH_SCALE = 3;
-	const int TILE_HEIGHT_SCALE = 3;  
+	const size_t TILE_WIDTH_SCALE = 3;
+	const size_t TILE_HEIGHT_SCALE = 3;
 
 	const double G = 9.80; //meters per second squared gravity
